# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DelhiGhar API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /visits:
    post:
      operationId: scheduleVisit
      tags: [visits]
      summary: Schedule a property visit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitInput"
      responses:
        "201":
          description: Visit scheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
  /visits/{id}/confirm:
    post:
      operationId: confirmVisit
      tags: [visits]
      summary: Owner confirms a scheduled visit
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Visit confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
...
  /reports:
    post:
      operationId: createReport
      tags: [reports]
      summary: File a fraud report against a listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Report filed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{listingId}:
    get:
      operationId: getReports
      tags: [reports]
      summary: Get fraud reports for a listing
      parameters:
        - in: path
          name: listingId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reports list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
...
components:
  schemas:
...
    VisitInput:
      type: object
      required: [listingId, buyerName, buyerPhone, visitDate, userId]
      properties:
        listingId:
          type: string
        buyerName:
          type: string
        buyerPhone:
          type: string
        visitDate:
          type: string
        userId:
          type: string
    Visit:
      type: object
      required: [id, listingId, buyerName, buyerPhone, visitDate, status, createdAt]
      properties:
        id:
          type: string
        listingId:
          type: string
        buyerName:
          type: string
        buyerPhone:
          type: string
        visitDate:
          type: string
        status:
          type: string
        createdAt:
          type: string
...
    ReportInput:
      type: object
      required: [listingId, reporterId, reasons]
      properties:
        listingId:
          type: string
        reporterId:
          type: string
        reasons:
          type: array
          items:
            type: string
        details:
          type: string
    Report:
      type: object
      required: [id, listingId, reporterId, reasons, createdAt]
      properties:
        id:
          type: string
        listingId:
          type: string
        reporterId:
          type: string
        reasons:
          type: array
          items:
            type: string
        details:
          type: ["string", "null"]
        createdAt:
          type: string
...