# lib/api-zod/src/generated/types/visit.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DelhiGhar API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Visit {
  id: string;
  listingId: string;
  buyerName: string;
  buyerPhone: string;
  visitDate: string;
  status: string;
  createdAt: string;
}